voluntariados-backend/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
gunicorn
python-dotenv
djangorestframework-simplejwt
django-allauth[socialaccount]
requests
cryptography